Allow using separate cpio archive for initramfs
This patch adds an option to initramfs filesystem target that allows the use of
a separate user-generated cpio archive in conjunction with a root filesystem
image.

The new option is called 'Initramfs cpio archive' and allows the choice between
two options:

- 'generate from root filesystem' is the default and behaves the same way as
  before: rootfs.cpio is generated by depending on cpio target.

- 'generate separately' creates a dummy rootfs.cpio target that is rigged to
  fail if there is no rootfs.cpio file in the images/ directory but otherwise
  does not try to generate one.

The second option is meant to be used with a post-build hook that creates a
custom rootfs.cpio.

Selection of these two modes is done using `choice` to allow for future
expansion as well as to avoid ambiguity about what 'the other' option would be.

Signed-off-by: Branko Vukelic <branko@outernet.is>