Allow compressed initramfs
In the previous patch that enabled separate initramfs image, only the
uncompressed image actually worked. This was probably also true of the origina
Buildroot implementation (unless I'm missing something crucial). This patch
provides a hackish workaround for using separately generated *and compressed*
cpio archive.

This patch also introduces LZ4 compression for filesystems in general so that
compression options are unified across the board.

Signed-off-by: Branko Vukelic <branko@outernet.is>